+++ packages/extension/src/background/index.js
@@ -5,3 +5,4 @@ export {
 } from './MessageHost.js';
 export { createBackgroundHost } from './createBackgroundHost.js';
 export { attachChromeRuntime } from './ChromeRuntimeAdapter.js';
+export { attachSessionMetaListener } from './sessionMeta.js';